finalizado crud de perfil

+++ resources/views/admin/plans/details/index.blade.php
@@ -15,31 +15,33 @@
         </div>
         <div class='card-body'>
             @include('admin.includes.alerts')
-            <table class='table table-condensed'>
-                <thead>
-                    <tr>
-                        <th>Nome</th>
-                        <th style='width:120px'>Ações</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    @foreach ($details as $detail)
-                    <tr>
-                        <td>{{ $detail->name }}</td>
-                        <td>
-                            <div class="btn-group">
-                                {!! Form::open(['route' => ['admin.plan.detail.destroy', $url, $detail->id], 'method' => 'delete', 'class' => 'form-delete']) !!}
-                                <button type='submit' class='btn btn-outline-danger'>
-                                    <i class='fas fa-trash'></i>
-                                    Deletar
-                                </button>
-                                {!! Form::close() !!}
-                            </div>
-                        </td>
-                    </tr>
-                    @endforeach
-                </tbody>
-            </table>
+            @if($details->count())
+                <table class='table table-condensed'>
+                    <thead>
+                        <tr>
+                            <th>Nome</th>
+                            <th style='width:120px'>Ações</th>
+                        </tr>
+                    </thead>
+                    <tbody>
+                        @foreach ($details as $detail)
+                        <tr>
+                            <td>{{ $detail->name }}</td>
+                            <td>
+                                <div class="btn-group">
+                                    {!! Form::open(['route' => ['admin.plan.detail.destroy', $url, $detail->id], 'method' => 'delete', 'class' => 'form-delete']) !!}
+                                    <button type='submit' class='btn btn-outline-danger'>
+                                        <i class='fas fa-trash'></i>
+                                        Deletar
+                                    </button>
+                                    {!! Form::close() !!}
+                                </div>
+                            </td>
+                        </tr>
+                        @endforeach
+                    </tbody>
+                </table>
+            @endif
         </div>
     </div>
 @stop
